Move UI abstractions and view model to adapter layer
