+++ OTAnalytics/adapter_ui/abstract_frame.py
