+++ OTAnalytics/plugin_ui/frame_canvas.py
@@ -6,12 +6,12 @@ import customtkinter
 from customtkinter import CTkCheckBox
 from PIL import ImageTk
 
+from OTAnalytics.adapter_ui.abstract_canvas import AbstractCanvas
+from OTAnalytics.adapter_ui.abstract_frame import AbstractTracksCanvas
+from OTAnalytics.adapter_ui.view_model import ViewModel
 from OTAnalytics.domain.track import TrackImage
-from OTAnalytics.plugin_ui.abstract_canvas import AbstractCanvas
-from OTAnalytics.plugin_ui.abstract_frame import AbstractTracksCanvas
 from OTAnalytics.plugin_ui.canvas_observer import CanvasObserver, EventHandler
 from OTAnalytics.plugin_ui.constants import PADX, STICKY
-from OTAnalytics.plugin_ui.view_model import ViewModel
 
 
 @dataclass
