+++ OTAnalytics/plugin_ui/frame_sections.py
@@ -3,9 +3,9 @@ from typing import Any, Optional
 
 from customtkinter import CTkButton, CTkFrame, CTkLabel
 
-from OTAnalytics.plugin_ui.abstract_treeview import AbstractTreeviewSections
+from OTAnalytics.adapter_ui.abstract_treeview import AbstractTreeviewSections
+from OTAnalytics.adapter_ui.view_model import ViewModel
 from OTAnalytics.plugin_ui.constants import PADX, PADY, STICKY
-from OTAnalytics.plugin_ui.view_model import ViewModel
 
 
 class FrameSections(CTkFrame):
